# specs/api-spec.yaml
openapi: 3.1.0
info:
  title: Mandarina API
  version: 0.1.0
  description: |
    Ingesta y consulta de Eventos normalizados (ver `CONTEXT.md` y `spec/mvp-fase1.md`).

    **Tiempo real:** el backend expone un WebSocket en `/ws`. Por cada Evento
    aceptado envía a todos los clientes un mensaje `LiveMessage`. La Actividad
    *Esperando* (ADR-0011) no tiene mensaje propio: los Eventos que la causan o
    la terminan llegan como `event.ingested` y el cliente refresca `GET /sessions`.
    El Uso de la suscripción (ADR-0012) sí lo tiene: `subscription.usage` con la última lectura de la cuenta.
servers:
  - url: http://127.0.0.1:4000
...
paths:
...
  /api/v1/events:
    post:
      tags: [events]
      operationId: ingestEvent
      summary: Ingiere un Evento enviado por un Adaptador.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '202':
          description: Evento aceptado, persistido y difundido.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: string }
        '400':
          description: El Evento no cumple el esquema.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [events]
      operationId: listEvents
      summary: Eventos más recientes primero.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: before
          in: query
          description: Devuelve solo Eventos anteriores al Evento con este id.
          schema: { type: string }
        - name: session_id
          in: query
          description: Solo Eventos de esta Sesión (AC-17).
          schema: { type: string, minLength: 1 }
        - name: event_type
          in: query
          description: Solo Eventos de estos Tipos; se repite para varios (AC-17, AC-21).
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/EventType' }
        - name: since
          in: query
          description: Solo Eventos recibidos desde esta fecha.
          schema: { type: string, format: date-time }
        - name: project
          in: query
          description: Solo Eventos de este Proyecto (AC-100).
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Página de Eventos.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Event' }
        '400':
          description: Parámetros inválidos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: Sesiones del board con su Estado y resumen (AC-15).
      description: Ordenadas por inicio, la más nueva primero; el orden no cambia con la actividad (AC-15).
      parameters:
        - name: since
          in: query
          description: Solo Sesiones con algún Evento recibido desde esta fecha.
          schema: { type: string, format: date-time }
        - name: state
          in: query
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/SessionState' }
        - name: directory
          in: query
          schema: { type: string, minLength: 1 }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Sesiones y valores disponibles para los filtros.
          content:
            application/json:
              schema:
                type: object
                required: [items, facets]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SessionSummary' }
                  facets:
                    description: Proyectos y Directorios de las Sesiones de `since`, antes de filtrar por Estado, Directorio o Proyecto.
                    type: object
                    required: [projects, directories]
                    properties:
                      projects: { type: array, items: { type: string } }
                      directories: { type: array, items: { type: string } }
        '400':
          description: Parámetros inválidos.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/sessions/{id}:
    get:
      tags: [sessions]
      operationId: getSessionDetail
      summary: Detalle de una Sesión (AC-18).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detalle de la Sesión.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionDetail' }
        '404':
          description: La Sesión no existe.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/test-runs:
    get:
      tags: [tests]
      operationId: listTestRuns
      summary: Ejecuciones de tests leídas de los Eventos `Bash` (AC-27).
      description: |
        Se derivan al consultar de los `tool.post` de `Bash` cuya salida trae el
        resumen de un runner reconocido (AC-26, ADR-0007). La más reciente
        primero; como máximo 500.
      parameters:
        - name: since
          in: query
          required: true
          description: Solo Ejecuciones recibidas desde esta fecha; la UI envía las 00:00 locales de hace 6 días.
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
        - name: kind
          in: query
          schema: { $ref: '#/components/schemas/TestKind' }
      responses:
        '200':
          description: Ejecuciones de tests y Proyectos disponibles para el filtro.
          content:
            application/json:
              schema:
                type: object
                required: [items, facets]
                properties:
                  items:
                    type: array
                    maxItems: 500
                    items: { $ref: '#/components/schemas/TestRun' }
                  facets:
                    description: Proyectos con alguna Ejecución desde `since`, antes de filtrar por Proyecto o Tipo de tests.
                    type: object
                    required: [projects]
                    properties:
                      projects: { type: array, items: { type: string } }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/skill-invocations:
    get:
      tags: [skills]
      operationId: listSkillInvocations
      summary: Invocaciones de skill leídas de los Eventos y de los Transcripts (AC-30).
      description: |
        Se derivan al consultar de los Eventos de la herramienta `Skill`, de los
        `prompt.submitted` que empiezan por `/nombre` y de las herramientas `Skill`
        de los Transcripts del agente principal y de sus Subagentes (AC-29). La más reciente
        primero; `items` trae como máximo 500, pero `stats` cuenta todas.
      parameters:
        - name: since
          in: query
          required: true
          description: Inicio de la ventana; la UI envía el inicio del periodo seleccionado (ahora − 1 h / 24 h / 7 d, o la época Unix para 'Todo').
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
        - name: session_id
          in: query
          description: Solo las invocaciones de esta Sesión (pestaña *Skills* del detalle).
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Invocaciones de skill, su agregado por Proyecto y skill, y Proyectos disponibles para el filtro.
          content:
            application/json:
              schema:
                type: object
                required: [items, stats, facets]
                properties:
                  items:
                    type: array
                    maxItems: 500
                    items: { $ref: '#/components/schemas/SkillInvocation' }
                  stats:
                    description: Una fila por Proyecto y skill con todas las invocaciones filtradas, ordenadas por total descendente.
                    type: array
                    items: { $ref: '#/components/schemas/SkillUsage' }
                  facets:
                    description: Proyectos con alguna invocación desde `since`, antes de filtrar por Proyecto o Sesión.
                    type: object
                    required: [projects]
                    properties:
                      projects: { type: array, items: { type: string } }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/subagents:
    get:
      tags: [subagents]
      operationId: listSubagents
      summary: Subagentes de todas las Sesiones (AC-35).
      description: |
        Se derivan al consultar de los Eventos y los Transcripts de las Sesiones
        con actividad desde `since` (AC-33). El más reciente primero; como
        máximo 500. La vista por Tipo de Subagente está en `/api/v1/agents` (AC-46).
      parameters:
        - name: since
          in: query
          required: true
          description: Inicio del periodo; cuenta un Subagente que empezó desde entonces o sigue en marcha. La UI envía el inicio del periodo seleccionado (época Unix para "Todo").
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
        - name: type
          in: query
          description: Tipo de Subagente.
          schema: { type: string, minLength: 1 }
        - name: include_internal
          in: query
          description: Incluir los Subagentes internos; por defecto se omiten.
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Subagentes y los valores disponibles para los filtros.
          content:
            application/json:
              schema:
                type: object
                required: [items, facets]
                properties:
                  items:
                    type: array
                    maxItems: 500
                    items: { $ref: '#/components/schemas/SubagentListItem' }
                  facets:
                    description: Proyectos y Tipos de Subagente del periodo, antes de filtrar por Proyecto o Tipo.
                    type: object
                    required: [projects, types]
                    properties:
                      projects: { type: array, items: { type: string } }
                      types: { type: array, items: { type: string } }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/mcp-invocations:
    get:
      tags: [mcp]
      operationId: listMcpInvocations
      summary: Invocaciones de Herramientas MCP y su uso por Servidor MCP (AC-42).
      description: |
        Se derivan al consultar de los Eventos de herramientas `mcp__…` y de las
        herramientas de recursos MCP (AC-41). La más reciente primero; `items`
        trae como máximo 500, pero `servers` cuenta todas.
      parameters:
        - name: since
          in: query
          required: true
          description: Inicio del periodo; la UI envía el inicio del periodo seleccionado (época Unix para "Todo").
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
        - name: server
          in: query
          description: Nombre del Servidor MCP.
          schema: { type: string, minLength: 1 }
        - name: session_id
          in: query
          description: Solo las de esta Sesión (pestaña *MCP* del detalle).
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Invocaciones, su uso por servidor y herramienta, las herramientas diferidas sin usar y los valores de los filtros.
          content:
            application/json:
              schema:
                type: object
                required: [items, servers, unused_deferred, facets]
                properties:
                  items:
                    type: array
                    maxItems: 500
                    items: { $ref: '#/components/schemas/McpInvocation' }
                  servers:
                    description: Una fila por Servidor MCP con todas las invocaciones filtradas, ordenadas por llamadas.
                    type: array
                    items: { $ref: '#/components/schemas/McpServerUsage' }
                  unused_deferred:
                    description: Herramientas MCP cargadas con `ToolSearch` y nunca invocadas en su Sesión.
                    type: array
                    items:
                      type: object
                      required: [session_id, tool_name, server, tool, loaded_at]
                      properties:
                        session_id: { type: string }
                        tool_name: { type: string }
                        server: { type: string }
                        tool: { type: string }
                        loaded_at: { type: string, format: date-time }
                  facets:
                    description: Proyectos y servidores del periodo, antes de filtrar.
                    type: object
                    required: [projects, servers]
                    properties:
                      projects: { type: array, items: { type: string } }
                      servers: { type: array, items: { type: string } }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/agents:
    get:
      tags: [agents]
      operationId: listAgentTypes
      summary: Comparativa de los Tipos de Subagente del periodo (AC-46).
      description: |
        Se deriva al consultar del ciclo de vida de los Subagentes (AC-33) de las
        Sesiones con actividad desde `since`; sin los internos. Ordenados por Lanzamientos.
      parameters:
        - name: since
          in: query
          required: true
          description: Inicio del periodo; cuenta un Lanzamiento empezado desde entonces o que sigue en marcha.
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Una fila por Tipo de Subagente y los Proyectos disponibles para el filtro.
          content:
            application/json:
              schema:
                type: object
                required: [items, facets]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/AgentTypeSummary' }
                  facets:
                    type: object
                    required: [projects]
                    properties:
                      projects: { type: array, items: { type: string } }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/agents/{type}:
    get:
      tags: [agents]
      operationId: getAgentProfile
      summary: Perfil de un Tipo de Subagente en el periodo (AC-46).
      parameters:
        - name: type
          in: path
          required: true
          description: Tipo de Subagente; `sin-tipo` agrupa los Lanzamientos sin Tipo conocido.
          schema: { type: string, minLength: 1 }
        - name: since
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: project
          in: query
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Perfil del Tipo; sin Lanzamientos en el periodo, con cifras a cero.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentProfile' }
        '400':
          description: Falta `since` o algún parámetro es inválido.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
...
  schemas:
...
    SessionState:
      description: Estado de la Sesión (AC-14).
      type: string
      enum: [active, idle, orphaned, closed]
...
    EventInput:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - harness
        - project
        - directory
        - session_id
        - event_type
        - native_event_type
        - occurred_at
        - payload
      properties:
        schema_version: { type: integer, const: 1 }
        harness: { type: string, minLength: 1, examples: [claude-code] }
        project: { type: string, minLength: 1 }
        directory: { type: string, minLength: 1 }
        session_id: { type: string, minLength: 1 }
        subagent_id: { type: [string, 'null'] }
        event_type: { $ref: '#/components/schemas/EventType' }
        native_event_type: { type: string, minLength: 1 }
        tool_name: { type: [string, 'null'] }
        occurred_at: { type: string, format: date-time }
        transcript_path: { type: [string, 'null'] }
        payload: { type: object }
        block:
          description: Solo en `tool.blocked`.
          oneOf:
            - { $ref: '#/components/schemas/Block' }
            - { type: 'null' }
...
    TestKind:
      description: Tipo de tests (ver `CONTEXT.md`).
      type: string
      enum: [unit, e2e]
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }